dom/tests/mochitest/cowl/test_labeled_ctx_response.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display"></p>
  <pre id="test">

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  var completedTests = 0;
  var passedTests = 0;

  function ok_wrapper(result, desc) {
    ok(result, desc);

    completedTests++;

    if (result) {
      passedTests++;
    }

    if (completedTests == 6) {
      is(passedTests, completedTests, "There are " + completedTests + " labeled objects tests that should pass");

      SimpleTest.finish();
    }
  }

  function labeled(iframe) {
    var wCWindow = SpecialPowers.wrap(iframe.contentWindow);
    //var wwIframe = SpecialPowers.Cu.waiveXrays(wIframe);

    ok_wrapper(wCWindow.COWL.confidentiality.equals(new Label("http://a.com/")), "Confidentiality label should be set according to ctx-confidentiality header");
    ok_wrapper(wCWindow.COWL.integrity.equals(new Label()), "Integrity label should be set according to ctx-integrity header");
    ok_wrapper(wCWindow.COWL.privilege.asLabel.equals(new Label()), "Privilege internal label should be set according to ctx-privilege header");
  }

  function labeled_self(iframe) {
    var wCWindow = SpecialPowers.wrap(iframe.contentWindow);

    ok_wrapper(wCWindow.COWL.confidentiality.equals(new Label("http://example.com/")), "ctx-confidentiality should evaluate to 'self'");
    ok_wrapper(wCWindow.COWL.integrity.equals(new Label()), "Integrity label should be set according to ctx-integrity header");
    ok_wrapper(wCWindow.COWL.privilege.asLabel.equals(new Label()), "Privilege internal label should be set according to ctx-privilege header");
  }

  </script>
  <iframe onload="labeled(this)" src="http://example.com/tests/dom/tests/mochitest/cowl/labeled.html"></iframe>
  <iframe onload="labeled_self(this)" src="http://example.com/tests/dom/tests/mochitest/cowl/labeled_self.html"></iframe>
</pre>
</body>
</html>
